You can now create tasks in specific projects
These projects must be specified in the config file under the
synergy-config project, with the name 'projects.yaml'. The
format should be:

  ---
  short_name_no_spaces: project_id
  other: id2

Where the ids are the numerical ids from liquidplanner.

This adds one command 'projects' which will dispaly all known projects in
private message (and a link to them in liquid planner).

To use the functionality:

  ++ some task #project_short_name
  >> someone: fix foo #project_short_name
  task for so and so: whatever #project_short_name

Note that if # is seen in the name, we assume everything after it is
a project to prevent mishaps. In this case, the project *must exist* or
we reject the creation.

New projects can be added by updating the synergy-config repo
and saying:

  synergy: reload projects