Fix targeting problem
Sometimes synergy would, at random, reply "does not compute" to some
person, if when she wasn't addressed. This would happen if someone else
happened to target her in some other channel at the same time.

This fixes that: the problem was that $1 wasn't being reset in an
unsuccessful match, so we'd think some unrelated message was, in fact,
targeted.

¡Perl!