Fix user loading from GitLab config
This is, I'm pretty sure, the source of the preference-init bug. In the
GitLab reactor, we load the cached users from state, but don't actually
hit the code that refreshes the user directory until the config is
loaded a minute later. (By which point, it's too late, because the
preference-setter says "oh we've already got a preference for this,
don't set it from config.")

So, temporarily, allow config to overwrite preferences.