Be case-sensitive about responding to “RFC”
I expect “rfc\d+” to be entirely false positives.

This is probably most notable in the MIME type “message/rfc822” which
crops up fairly often in the email space. If this approach of requiring
capital “RFC” is considered too radical, an alternative would be to
specifically not respond to “rfc822” (case sensitive). But that’s more
work to implement.