HTTPEndpoint: role for HTTP endpoint on a channel or reactor
Adds a role that automatically hooks up an endpoint to a channel or
reactor, including authentication if the config demands it.

The base HTTPServer remains available for anything that wants to do
something special; in that case, they're on their own and will have to
handle auth and anything else directly.

I've changed HTTPServer to call the apps with the PSGI env, not
Plack::Request. That's the PSGI way anyway, but also makes the wrappers
much simpler because we don't need to keep unwrapping and rewrapping the
request for things that expect an env.