Support for deleting users
This is a bit half-baked at the moment, but you can set "deleted: 1" in
the user directory. Deleted users won't resolve, so most of the things
you want to do with them stop working. This still does some silly things
like keep state around for deleted users, but that seems fine for now.