Slack: move the frame fiddling into External class
I'm pretty happy with this commit! It removes the need to plumb the
frame_id up through the Event and back down into the Channel. Instead,
Channel::reply should return a Future, and note_error can just do stuff
when that future finishes.

In External::Slack, every time we send a frame (which is, in practice,
either a ping or a plain_text frame), we'll keep a success future and a
timeout future for that frame. When we get a response to that frame, we
cancel the timeout future and mark the success as done (with the
contents of the actual frame sent, so that other places keeping track of
the future get the reasonable data).  If we never get a reply to that
frame, the timeout will fire, triggering a reconnection to slack. (This,
conveniently, eliminates the need for the pong timer and its
machinations.)

Overall, this makes the logic much easier to follow, and keeps the frame
manipulation down in the low-level Slack connection class, where it
rightfully belongs. It also means that the main Slack channel doesn't
need to muck around as much with the External::Slack's internal state
(which it needed to do with the pong_timers and frame tracking).