Give Synergy::Event a way to mark a pending reply
In the futuristic near future, our HTTP requests will be fully
asynchronous, so we can't depend necessarily depend on an event being
marked handled by the time control flow reaches the default case in the
hub.

For instance: you have a listener that makes a bunch of HTTP requests,
gathers them into a reply, and maybe responds. We don't want to mark it
handled unconditionally, because it might not respond and maybe that's
ok (e.g., a mention-in-passing listener). But if it's targeted, we might
still want to mark that as an error. In that case, you can set
$event->pending_future to a future, and if that future fails, we'll send
the "Does not compute" response.