Yelling: yell harder, yell louder
This adds a hook into $event->reply just before sending the message. If
there are any pre_message_hooks defined on the $event's from_channel, it
will run those in turn, passing in a reference to the plain text and the
alts (along with the event), so that they can take action on those if
they want.

This is a bit of a hack, but is more or less the right way to do it.