Slack: don't double-encode attachments
When we switched from sending form-encoded data to JSON over the wire to
Slack, we (I) broke attachments. When we were sending form-encoded data,
attachments had to be sent as JSON-encoded strings. We *were* encoding
those correctly, but when we switched to sending JSON, that meant we
were double-encoding the attachment data. Fortunately, we now just send
JSON data instead of strings of JSON!