Bump Net::Async::HTTP version for SNI support
Really, that was added in 0.43, but 0.44 can't hurt, right?

This seems to fix issues we're presently having (out of nowhere!)
connecting to slack.com with 0.42 which will occasionally fail
with:

  SSL connect attempt failed error:14094410:SSL routines:ssl3_read_bytes:sslv3 alert handshake failure failed [ssl] at lib/Synergy/External/Slack.pm line 122

If I downgrade my version to 0.42, I can reproduce the failure
occasionally, but bumping it up to 0.44 I can't so... yay?