Damage report - don't reuse last 'report for ..' user !
I've added two protections here, one is to make the capture
for "report for <user>" optional which causes $+{who} to be
properly reset if there was no user specified. The other is
to not even use $+{who} at all if we don't match for some
reason. This way, we don't reuse stale capture variables.