Slack: don't repond to link expansions
When Slack auto-previews a link, it sends a message_changed event.  If
you said "++ <link>", this would often lead to Synergy saying "hey, I
can't respond to that," because she'd respond and then get the
message_changed event when Slack expanded the URL.

This fix probably isn't perfect, but it seems right. If a message didn't
used to have attachments, and it was edited to have them, we'll just
ignore that edit, since it likely didn't come from the user. (I don't
*think* users can add attachments to messages anyway...at least not to
messages that would otherwise be directed at Synergy.)