slack: do not respond to "synergee"
This is a dumb hack: when I converted synergy to a Slack app, I gave her
perms to add a user with the name "synergee" because I thought "synergy"
would conflict. So now "synergee ++ do a thing" works, which is not
ideal, since we use often that as a way of joking about making tasks
we'd never do. I *think* that reinstalling the app to our workspace
would fix this, but I'm not entirely sure and I don't want to make
everyone open yet another DM with synergy, so here we are.