UserDirectory: be better about deleted users
Before, $directory->users would return you only active users, but the
delegations like ->user_named would pull deleted ones. This was
confusing because we have a deleted user named 'chris' and an active one
with a nickname 'chris'.

So now, we'll keep a reference to just the active ones in the directory,
and move the delegations to it. Any time you set a user, that'll go
through the real one, but doing so will clear the active users so it
should always be correct.