TimeClock: add an evening report
First, we get rid of the "user started recently" method, and instead
give users a method to provide their shift hours, if any, for a given
day.  Among other things, this lets us tie into the availability checker
so we can decline to send users any timeclock reports if they're off
today.

Then we send an evening report to a user if they're going to go off the
clock in the next half hour.

Note that this system is currently built on the idea that…

1. a user only works one set of continuous hours in a day
2. a user's shift never spans a date boundary
3. shifts are never really really short