slack: add method for form-encoded calls; use it
Sometimes, in group DMs in which Synergy was a party, you'd send a
message and get a cryptic reply like "my prom reactor crashed." That's
because prom would call describe_conversation on the group message, and
not find the ID, and try reloading them all from Slack, which would
sometimes get rate-limited.

The reason we could never find them is that when we loaded group
conversations, we passed `{"type": "mpim"}` to the conversations.list
endpoint, but that endpoint doesn't accept JSON (!), so it was ignoring
the argument and giving us all the public channels instead. That meant
that we'd never actually load the group conversations at all!

The dumb fix is to add another arg to api_call, so you can request
`form_encoded => 1` to make the call with
`application/x-www-form-urlencoded` instead of `application/json`.