VO: fix output bug when already in maint
This morning I started a maintenance mode and we already happened to be
in one, and I got two responses:

1. VO already in maint!
2. VO now in maint! HASH(0x562d06ce0080) alerts acked! Good luck!

That's because if we were already in maint, we'd return $event->reply
from the branch, which would resolve to an empty successful future, and
the ->then would get "not enough arguments for subroutine" because it
was expecting the result from ->_ack_all.

The fix is straightforward: just use some sentinel value. (Incidentally,
this was caught by the test, which was failing and is no longer!)