commando: refactor predicate building a bit
If we're given an explicit match, we must not build our names into the
matcher. That's because, at least, the VO reactor uses 'maint start' to
start maint but /^maint$/ to query the status. If we always add our
name, then 'maint' will also match 'maint start', because the predicate
we build is effectively /^$name/.