inabox: fix bananas error case with needs_all
In testing, sometimes I would see *super* strange errors deep down in
IO::Async if part of needs_all failed. I *think* (but cannot say for
sure) that this happens because both HTTP requests are to the same
endpoint and might be reusing a connection, so when the first fails
IO::Async closes the connection, causing the second one to fail. This
would throw "Spurious read while idle" from Net::Async::Connection.
I spent an inordinately long time tracking this down, and it doesn't
seem easy to fix, so I will just do these calls in sequence instead.

On the bright side, this fix lets the test run in parallel again!