VO: manage a Slack usergroup
This copies Bort's code for doing the same (which has now been removed).
This isn't ideal because the slack method to update user groups requires
a privileged API token; this commit also includes the plumbing for that.

Unlike Bort, this checks oncall on a timer, every 2.5 minutes. That's
because the current VO API for checking oncall no longer includes the
next scheduled shift change, so figuring out when it changes
is...non-trivial, and involves munging a bunch of schedules and stuff.