VO: when demaint/resolving, resolve before demaint
Before, there was a race: if you exit maint, VO immediately sends phone
alerts for everything that's active, even if they're going to be
resolved in a quarter-second. That's really annoying, so now we resolve
everything, then tell VO to stop maint, at which point there shouldn't
be anything left to alert.