LP: add nagging-hours preference
This is now separate from business hours (but falls back to business
hours, if nagging hours is unset). This is so that I can have business
hours that reflect the time I actually work (so other humans know) but
not be nagged in the first half-hour of work, in case I start a bit late
or whatever.