GitLab: eliminate custom repo shortcuts file
This is a vestige of the time before we had user prefs, and so we stored
preferences in a special repository in GitLab. We don't need to do that
any more, so we can mostly get rid of this behavior.

This also removes the reload_repos command, because there's nothing left
to reload. Custom config will be in synergy's config file, so a restart
will pick it up.

(This is in service of killing off the config repo.)