Slack: rework edit-reply logic and data structures
Before, we stored data on messages we replied to as a hash of
$timestamp => $reply_data. This was buggy, because we sometimes reply
more than once to a single message, and whatever our last response was
wound up overwriting the previous ones.

This had at least one side effect that we noticed all the time. An
example: user says something like "synergy >> pat fix thing, ptn 12345",
attempting to create a task for pat in reference to a particular Zendesk
ticket. We respond to that message twice: once for the LP task creation,
once with a link to the Zendesk ticket, and the link reply data would
win. This is a problem if the LP task creation causes an error (if
synergy doesn't know who Pat is, for example), because the first reply
is marked was_error, but the Zendesk link expando is not. That means if
the user tried to edit the message (which ordinarily, they'd be able to
do), Synergy would say "hey, that message didn't cause an error, I can't
deal with your edit, sorry."

My first pass at fixing this was to keep the reply data the same, and
set was_error as `$existing_reply->{was_error} || $new->{was_error}`.
That worked, at least to fix the immediate problem, but was inelegant:
when synergy replies to an error, she deletes her original error reply
(the one that says something like "hmm, that didn't work"), but with
this fix, she'd always delete the last one; in the example above, you'd
be left with "hmm, that didn't work" and she'd delete the Zendesk link.
That kinda stinks.

The actual fix is just: for every message we respond to, keep a list of
replies instead. This requires slightly more bookkeeping, but when an
edit comes in, we can noticet that only part of it caused an error and
delete just that part of our response.