PagerDuty: implement resolve behavior
This simplifies the resolve behavior a bit, which we can do because I'm
choosing not to port demaint /resolve. That's mostly because it's
unnecessary, because PagerDuty (and now, incidentally, VictorOps) do
not even send alerts during maint, so there's not really anything to
resolve. VO does "resolve everything that fired in the 10 minutes before
maint started," but we found this behavior confusing and difficult to
keep straight.