PagerDuty: add API token pref and use it
This diff is slightly tedious (sorry!). We want to be able to tell
PagerDuty _who_ is making certain kinds of calls (acking, resolving,
starting/ending maint or whatever), which means we need to have an API
token. This adds that preference, and copies the logic from the LP
reactor so that setting your API token automagically sets your PagerDuty
user id as well.

This commit also adds a new API request method for using the users's own
API key (`pd_request_for_user`), and uses it for all the non-GET
requests. I thought about having _everything_ use this, which didn't
quite work for two reasons:

- passing through the user through all the abstractions kinda stunk, but
  more importantly...
- we still need to have an API token for doing the system-automatic
  things where we don't have an event (i.e., an instigating user):
  updating the oncall group and checking for long maint

If it were just the first, I'd deal with the tedium, but since we need
an API key _anyway_, we'll normally use that one, and only use the
user's API key when we need to.