PD: cope better with failure to update oncall list
This was announcing every two minutes because the entirety of the oncall
list hadn't yet set up their PagerDuty user ids, which meant that when
we tried to convert it to a Slack list, we got the empty list. This
caused the slack update-at-oncall API call to fail, which meant that we
never updated our cache, which in turn meant that we announced a change
every time the timer ticked.

There are two fixes here:

- if we don't get any valid slack users, bail early
- don't run the oncall announcement unless we successfully update the
  slack group
- (some annoying formatting tweaks I didn't want to hassle Andrew about
  during code review)