rototron: do not die on bad JMAP requests
Recently, Fastmail was down (sigh). This caused synergy to crash over
and over again, because _leave_days does a JMAP request and then calls
`sentence_named` on it. If that result is a Failure object, there is no
`sentence_named` method and we crash.

This commit copies what we do elsewhere: the JMAP call is done in an
eval, so inside that eval we call ->assert_successful, which will die on
failure and then skip the bits after that try to get useful information
out of the JMAP response.