PagerDuty: tweak format of incident summary
This required a bit of finagling, because we want to pass both rich text
(for slack) and plain text (for slack fallback and non-slack channels).
The important change here is to make _active_incidents_summary return a
hashref (with 'text' and 'slack' keys), and then updating the things
that care about it to match.