Hub: change how we avoid hub component name collision
The previous code used a single global hash to track channel and reactor
registration.  By "global", I really mean it.  It was a top-level
lexical variable, meaning that two Synergy instances in one process
would block each other from re-using component names!  While this isn't
likely to really happen, it's a circumstance we always keep in mind.

Instead, we just check every name on each registration.  It's not
efficient, but it's cheap, and happens rarely.