Vestaboard: overhaul to use a single PSGI application
The previous HTTP architecture was really goofy, because until now, the
Synergy HTTP server required that each path had a distinct application.
Now, instead, we use Plack's URLMap to register whole subtrees of path
space.  That means this change is pretty significant!

First, I extracted the POST handling, just to cut down on the complexity
of the Plack app sub, which was going to get more complex as it absorbed
other behaviors.

Then I folded in the "secret URL component" bits, which map the space
under /${secret} to fetch the current board content or redirect to an
image server.  Those were the "content server".

Then I folded in the "asset servers", which look for files on disk in
the asset path.  This went from one stupid thing (find all files,
register single-file servers) to another stupid thing (try to safely
find files inside a path using Path::Tiny, 404 if not there).  I believe
this is all correct.