Help: make help_entries an optional method
The comment explains it, but: requiring that the reactor supply help and
that roles decorate it leads to complication when one wants to replace
the help entries method wholesale, because of Moose composition order
and conflict resolution.  Instead, this commit makes help_entries a
method that a reactor may choose to implement or not.