CommandPost: a new framework for coding reactions
In previous versions of Synergy, all reaction was governed by Listener
objects, which set the form for how reaction could occur.  This was
replaced by the "potential_reactions_to" method, which more readily
allowed reactors to choose their own means by which to react to events.

CommandPost is proof of concept of a new way to build reactors.  It puts
more things in one place.  It simplifies the common cases of writing
named commands and mention-listeners.  It helps avoid duplication, for
example repeating a regex in the reaction body as well as the
EasyListening-style predicate sub.

Instead of predicates, CommandPost provides "matchers", which return a
value that can then be fed to the reaction method, allowing for
single-pass parsing (for example).

In addition to providing an implementation of "potential_reactions_to",
CommandPost provides four declarative helpers:

* command NAME => {ARG} => METHOD

This declares a command: a reaction that must be targeted and exclusive,
and which responds when a message starts with the command name.

* listener NAME => METHOD

This declares a listener: a non-exclusive reaction that gets called on
every event.

* reaction NAME => {ARG} => METHOD

This declares a flexible reaction that may be target and may be
exclusive.  It's similar to the listener declarations in EasyListening.

* help TITLE => TEXT

This adds a help entry to the reactor, in case you want to add help
topics that don't correspond to a command.