CommandPost: a big refactor, split into two modules
We now have two libraries:
*  Synergy::Role::Reactor::CommandPost -- a reactor with a command post
*  Synergy::CommandPost -- helpers for setting up your command post

(That second one, maybe, should get renamed to have ::Util:: in it.)

This is what both Michael and I had preferred, but my early attempts
couldn't quite get it right.  The more I worked on making this easier to
test, though, the more I circled in on the design in this commit.

This code is of particular importance:

    my $commandpost_override = $arg->{commandpost}; # Used for testing.
    my $get_cmdpost = $commandpost_override
                    ? sub { $commandpost_override }
                    : sub { caller(1)->_commandpost };

This, in the helper module's importer code, says that *normally* the
helper will deal with the helper-importer's command post, but lets you
write:

    use Synergy::CommandPost -default => { commandpost => $obj };

...to re-use an object for which you have a handle.  While this might be
useful in a reactor, it's definitely useful in testing, as you'll see in
the next few commits.