CommandPost: add parsers, tweak matchers
A "matcher" is a reaction property.  It's a callback that takes the args
($event_text, $text) and returns either undef, meaning "do not match",
or an arrayref, meaning "call the reaction handler, passing the elements
of the referenced array as the arguments."  This delivers on part of
CommandPost's promise to improve on EasyListening, by allowing code to
be deduplicated between the predicate and method.

A "parser" is very similar, but is part of a command rather than a
reaction.  By the time a command is being called, we've committed to
providing a reaction.  The parser can either return an arrayref, the
elements of which will be arguments to the handler, or it can return
undef, in which case no args are passed after the event.

All this is a bit experimental now...