Linear: implement issue identifier listener
This is heavily based on work by Rasha, but I had to rework it
significantly to apply it on the just-merged commandpost-extra branch.
In doing so, I also made some changes to its operation for a few
reasons.

Rather than apply this as Rasha and risk her getting the blame for some
complex screw-up introduced, I am adding this comment to allow her to
take the credit for overcoming the friction on this task while admitting
that any bug left in here is probably my fault. ;-)