Hub: fix logic in "does not compute" response
I added this in 1f9d418b, but the logic was buggy, and it would only use
the environment's `wtf_replies` setting when `$event->from_user` was
false, rather than if the from user had no wtf_replies.

In fixing this, I also realized that I'd forgotten the Array trait on
the environment's attr, so the delegation crashed anyway, once we were
calling it. Silly michael!