PagerDuty: implement oncall overrides
We want this for a workflow where there is one person oncall, and that
person needs to go to the store or whatever. This implements an
extremely limited surface for doing so: it only works for periods of
less than 8 hours, and only if one person is oncall.  Both `give oncall
to WHO for DURATION` and  `oncall override WHO DURATION` work, and do
the same thing.

We might expand this in the future, but this is a start.