dist.ini: fix dupe handling of Makefile.PL
We have to exclude Makefile.PL from Git::GatherDir, because we're
copying it from the build. This worked the first time because it didn't
exist, but failed the second time around, and we don't run dzil build on
a regular basis.

(Really I should have just used PluginBundle::MMCCLIMON instead.)