Preferences: convert to async describers
This paves the way for some future improvements, when describers might
want to do more interesting things like making HTTP requests (the Linear
reactor will probably want this). Now, describers are expected to return
a future that resolves to a string description. For now (and maybe,
forever!) we call Future->wrap, so they can just return a bare string in
the easy case.

This also requires some reworking in the prefs reactor, because we must
wait on all the describer futures before formatting the list of
preferences.