Rototron: attempt to fix a Monday time zone bug
Australians report that their morning reports do not contain duty
rotations; we're pretty sure this is because they're sent on Sunday UTC,
when there is no duty assigned. I think we can fix this by always
passing in a time zone to current_officers_for_duty, which will affect
the calendar search we're doing.

I'm not totally sure this will work, because time zones are weird and
the Rototron code is also kinda weird, but I don't think it's likely to
be _worse_.