linear: allow passing /current to 'agenda'
Sometimes people only want to see the current things on the agenda, and
not things they've planned for a few weeks out. This is a quick and
dirty way of doing that; really I should put this into Linear::Client's
search helper as currentCycleOnly => 1 or something, but this is
expedient for now.