Slack, Discord: fix detection of event targeting
The use of !!$1 was left over from a previous version of the code.  The
tests for the common code didn't cover this because the broken code was
only in the real webservice-based reactors, not the test one.  Likely
meaning: we should put more of this code into the common code!