InABox: convert "handle_box" to async/await
The eval of the handler is not great, as we lose the reporting of
user-visible error messages.  We should fix this generically in Synergy,
because the previous scheme is not sustainable.

I am surprised we never had bigger problems.  The signature on the
->then-handling sub would have failed to handle Futures that failed with
a single value, for example.

Meantime, this prevents broken async/await handler code from having its
exceptions lost in the ether or turned into signature arity violations!