InABox: remove unneeded magic incantation
I can no longer reproduce the problem that led to this line being added.
Meanwhile, there were a number of mystifying errors that had come from
an ->else handler on a sequenced future having a strict arity, which is
not compatible with the block being called with a 0+ sized list of items
packed into a failed Future.

Removing that code made things much, much less mysterious.