Hub: all Reactor and Channels now have "readiness"
It is useful to schedule things to happen no sooner than when a given
channel, or all channels, have come online.  The "start" method is meant
to do some of this work, but often start methods send work into the
background.  If we were better about tracking that, we could start many
"start" methods up and know when they were all or each done.

That's what this commit is for.

Now, every Reactor and Channel has a "readiness" attribute, which is a
Future that starts life pending.  It's meant to become done by something
calling and waiting on the component's "become_ready" method, which just
calls "start" and then marks the component ready.

Just like before, the Hub will ensure that the start methods are called,
but it now does it by awaiting become_ready, via a Future->wait_all,
first on all reactors, then on all channels.

Most things were converted to just return an empty "done" future, right
now, but we can return something more meaningful easily now.  For
example, the Slack and Discord reactors now return a future based on the
old "is_ready" behavior on their External helpers.  The Slack and
Discord reactors are ready when their External is ready.

The upshot of this is that a Reactor can now, during its own start
method, write something like this:

  my $target_channel = $self->hub->channel_named($self->target_channel);
  $target_channel->readiness->on_done(sub {
    ... send message here ...
  });

There won't be a race between Reactor and Channel startup, because they
will cooperatively sequence.  (You can't use "await" on that Channel
readiness, though, because channels are meant to wait on reactors!
Don't create a deadlock!)