Slack, Discord: do not await readiness
This was not quite correct.  First off, the ->readiness on the external
needed to be await-ed, since it's a future in an async sub.

Secondly the readiness properties there aren't actually sequenced
futures, they're just predicates that check on action completion.  We
need some kind of loop here, with a yield-to-next-cycle.  Instead of
doing that now, I'm just going to not await those.  It will be fine
until I do what's needed: properly make the readiness on the externals
track actual readiness without needing to be called repeatedly.