Slack: eliminate "setup" method
We don't need to call setup.  Also, it was busted.

Setup was there to kick off the loading of lookup tables (user, dm
channels, etc.) from Slack's API, which would finish later.  Those are
now triggered by awaiting the External's `readiness`.  We don't need to
wait for a `hello` frame, because the RTM interface is on WSS, and the
lookup table fetching is over HTTPS.  They can operate independently.

*Also*, the refactoring of the load_X methods into async methods meant
that they now return futures that need to be awaited, but the setup
method did not retain them.  Things that await the ->readiness method
work, though!  And now that's the only code path.