Slack: block on connection when sending a frame
Right now, if we fail to send a frame in a timely fashion, we reconnect.

The reconnection was calling a newly-async sub, but was not awaiting its
result.

The whole stack needs to be properly async, but for now we can just
block.  This will only come up when Slack is busted, so this blocking is
not the end of the world.

But not great.