InABox: after creation of a box, ssh to it for setup
This is the first in an n-step process to allow Synergy to manage
post-creation execution of work on the box.  (DigitalOcean has
provisions for something like this using cloud init, but we don't use
that.)

If Synergy has been given an ssh key to use, then when she's finished
creating a box, she'll use that key to connect to its ssh and run a
setup command.

Right now, the setup command gets no arguments and runs unconditionally
upon box creation.  We'll make this optional and separable in a future
commit, along with adding preferences and the ability to pass arguments
to the setup command.