GitLab: simplify the MR report
The MR report can now be defined in terms of a normal search, relying on
the normal search and pagination code.  We now only fetch up to 26
items, and will say 25+ if we're over 25.  Isn't it enough to know that
you're *that* underwater, after all?

Possibly this can be simplified if we can avoid post-fetch filtering.
This might be possible since it seems that "search by is-approved" is
now part of the standard GitLab API, not only the ultra premium
offering!