GitLab: add "maybefor" filter and use it
Say you want to know what code you are probably meant to still be
working on.  That's:

* author: you
* assignee: you OR NOBODY
* approved: no

The problem is the "or nobody", as indicated by the capital letters.
You can't query for this in the merge request API, so we do it page by
page in Synergy.  Unfortunately, the links we provide into GitLab are
then not quite right, but I don't think we have a better options.