InABox: add a sleep-like loop to tests
The comment added in this commit describes things, so I will reproduce
it here:

This is sort of idiotic, by which I mean a few things:

* This code is doing something stupid because that's how to make it
  work.
* I am (only a little bit) an idiot for making this necessary.
* This is stupidly unoptimized because optimizing something like this is
  stupid.

The above *should* work, but it only measures that every event has been
sent through the queue.  Instead, we should make exhaustion on the test
channel mean that every fired reaction handler has fully completed.
This is probably doable, but it will need some real thinking.  Instead,
we are going to just let the loop run for one quarter second.

Annoyingly, this code was not needed on Rik's laptop, either because
it's very fast or because the loop behavior is somewhat different there.

Reverting this commit later will be a delight.