InABox: use event completeness to test better and faster
This reverts commit 06d7989e6df328c4f568e5bbd1d9e7f6efb48bd5 (that's
"InABox: add a sleep-like loop to tests"), but it does much more.

Now that we can tell whether an event's handling is complete, we can
rely on that for knowing that all events have been fully handled.  That
seems to have been the previous purpose of the "wait" commands in here.
That is, when we expect:

    X.AMPLE: box create
    SYNERGY: [+react: working on it]
    SYNERGY: [-react: working on it]
    SYNERGY: It's done!

Previously, testing might complete after the first reply, when the other
replies were still sitting in queue.  The test loop's wait_for would see
that the input message queue was exhausted and stop testing.  That's
why, I think, we had these very long waits.  (Why six seconds?  I'm not
sure!)

The new event completeness property will not be done until the final
reply from the reaction code, so the test loop will keep running until
after the third reply.  This should eliminate any need to wait.

While I was in here, I've also rewritten a bunch of the test code to be
a bit more helpfully abstracted.  You can say "given this command,
expect these replies".  The old code did that too, but repeated a bunch
of code to do it.  The new code is clearer (says me), and it is
definitely shorter.  (And did I mention it's faster, too?)