Event handling: track whether an event is complete
"Complete", here, means that every potential reaction has returned a
future that has become done.  The Event class gets a "completeness"
attribute that is a Future that begins life pending.  It will be set to
done when all PotentialReaction futures are done.  (The name
"completeness" isn't great, but got the job done for now.)

Doing this means that we could, in one reaction, attach a callback to an
event to be run when the event has been fully handled.

More importantly, this should enable the Test channel to complete a test
program *not* immediately after it thinks it finished running every step
in the program, but instead also after every event injected is complete.
This should help avoid some races.

Finally, we keep track, in the Hub, of every event in flight and all its
pending reactions.  This should make it easy to see when more reactions
have attached to an event than expected, or when an event lingers
incomplete for an unexpectedly long period.