Tester: allow test programs on existing hubs
The Tester code helpfully provides `testergize`, a one-stop "create a
Synergy, run a test program, and get a Tester::Result" interface.  I
want a way to run a test program on an existing Synergy, though.  This
commit makes that possible by adding a `test_synergy` method that takes
an existing Synergy (and, optionally, a logger).  Then, `testergize` is
rewritten in terms of `test_synergy`.