Tester: refactor to make a Synergy::Tester::Hub class
The old interface was weird.  Tester could make a new hub and test with
it, or it could test with a hub you pass in.  You ended up calling
static methods and passing in data to use.  The logger behavior was
twitchy.  The whole thing was just a bit complicated, probably just
because of accreted maldesign.

The new design is much simpler OO sort of stuff.  `testergize` is gone.
You can call a static method on Synergy::Tester to get a new hub.  That
method is a specialized constructor that builds a Synergy::Tester::Hub,
which is a subclass of the normal Hub with one new method.  The new
method runs a test program and returns a Result.

This means no more running a no-program testergize to get a result just
to extract the hub to then run a program after further setup.  It just
makes sense.